refacto: deprecate assistant_builder/
