+++ front/components/agent_builder/AgentBuilder.tsx
@@ -29,12 +29,14 @@ import {
   transformDuplicateAgentToFormData,
   transformTemplateToFormData,
 } from "@app/components/agent_builder/transformAgentConfiguration";
-import type { AgentBuilderAction } from "@app/components/agent_builder/types";
+import type {
+  AgentBuilderAction,
+  AssistantBuilderMCPConfigurationWithId,
+} from "@app/components/agent_builder/types";
 import { ConversationSidePanelProvider } from "@app/components/assistant/conversation/ConversationSidePanelContext";
-import type { AssistantBuilderMCPConfigurationWithId } from "@app/components/assistant_builder/types";
-import { useNavigationLock } from "@app/components/assistant_builder/useNavigationLock";
 import { appLayoutBack } from "@app/components/sparkle/AppContentLayout";
 import { FormProvider } from "@app/components/sparkle/FormProvider";
+import { useNavigationLock } from "@app/hooks/useNavigationLock";
 import { useSendNotification } from "@app/hooks/useNotification";
 import type { AdditionalConfigurationType } from "@app/lib/models/assistant/actions/mcp";
 import { useAgentConfigurationActions } from "@app/lib/swr/actions";
