+++ front/components/agent_builder/shared/tables.ts
@@ -102,3 +102,4 @@ export async function expandFoldersToTables(
 
   return tablesArrays.flat();
 }
+
