+++ front/components/assistant/TagsSuggestDialog.tsx
@@ -168,3 +168,4 @@ export const TagsSuggestDialog = ({
     </Dialog>
   );
 };
+
