+++ front/components/data_source_view/TagSearchInput.tsx
@@ -101,3 +101,4 @@ export const TagSearchInput = ({
     </div>
   );
 };
+
