+++ front/components/spaces/SpaceSelector.tsx
@@ -157,3 +157,4 @@ export function SpaceSelector({
     </>
   );
 }
+
