+++ front/lib/swr/actions.ts
@@ -2,7 +2,7 @@ import uniqueId from "lodash/uniqueId";
 import { useMemo } from "react";
 import type { Fetcher } from "swr";
 
-import type { AssistantBuilderMCPConfigurationWithId } from "@app/components/assistant_builder/types";
+import type { AssistantBuilderMCPConfigurationWithId } from "@app/components/agent_builder/types";
 import { emptyArray, fetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
 import type { GetActionsResponseBody } from "@app/pages/api/w/[wId]/builder/assistants/[aId]/actions";
